refactor(kubernetes): Remove V1 dependence from V2 test
The V2-only KubernetesAccountResolverSpec depends on the
V1 credentials object KubernetesV1Credentials.  It's not actually
using any properties specific to the V1 credentials; this class
exists only to ensure that the resolver only returns V2 credentials.
Replace a mock for the V1 credentials with a mock for the interface,
which equivalently exercises the test case that we don't return
KubernetesCredentials that aren't KubernetesV2Credentials.