refactor(kubernetes): Reduce API surface of KubernetesKindProperties
Only the statically-defined kinds are allowed to set isDynamic to false,
or hasClusterRelationship to true. In order to make this statically
true, make the full constructor private and expose a factory method
that doesn't allow those properties to be set for use outside the class.

(Some of the usages of the constructor in tests *did* set isDynamic and
hasClusterRelationship, but these were just testing the API that was
available; now that these are not exposed outside the class the tests
can just use the factory method.)