refactor(kubernetes): Simplify getRegisteredKinds
The only time we ever use the getRegisteredKinds method in
KubernetesKindRegistry, what we actually want is a list of
the global kinds (so we can set up the core caching agent).

We do this by getting all the kinds and filtering on the
isDynamic property; it would be simpler to just return the
global kinds in the first place. A kind is global if and only
if it is not dynamic.