refactor(kubernetes): Remove useless call to register kind
On account creation, we iterate over all kinds in the account
and call registerKind. Since we only have the name of the kind
and no other information, we just assume the kind is namespaced
and pass true.

Since the kind registry will return a set of default properties
(including isNamespaced set to true) for something not in the
registry, skip the needless call to register these kinds with
default properties.