refactor(kubernetes): Pass custom resources to registry constructor
Each account has a set of custom resources that are statically
defined in the account config. We register each of these into
the KubernetesKindRegistry, but currently do so in a somewhat
opaque way, by depending on a side-effect of the
KubernetesResourceProperties.fromCustomResource function we use
when creating a different property on the account.

Remove the side effect from the fromCustomResource function and
instead pass the list of custom resources to the constructor for
KubernetesKindRegistry, allowing it to register resources there.