refactor(kubernetes): Allow KubernetesKindRegistry to look up CRDs
To avoid needing to pre-register all CRDs into the kind registry, the
registry should be able to look up kind information in the cluster if
it comes across a kind that it does not have any information about.

When creating a KubernetesKindRegistry, the caller can provide a
Function<KubernetesKind, Optional<KubernetesKindProperties>>; when
looking up a kind that is not in the registry, the registry will
call that function to try to resolve the kind properties. If the
function returns a result, the properties are recorded in the registry
(so subsequent lookups of the same kind don't call the function again).
If the function returns an empty Optional, we return a set of default
properties as before.

This commit does not actually use the new functionality; this will
come in a later commit.