refactor(kubernetes): Look up CRD details on-demand
Create a new function getCrdProperties that on-demand looks
up the properties of a CRD and pass this to the constructor
for the kind registry. This removes the need to pre-register
kinds that are fround when listing CRDs as the registry will
look up any kinds it needs, and also removes the last remaining
side-effect of calling liveCrdSupplier.

The only call to getCrds is from the unregistered CRD caching
agent. Given that this happens once per caching cycle (along
with a lot of other lookups) there's no real benefit to
memoizing the result with a timed expiration. Remove the
memoization and just have getCrds do a live lookup.