fix(kubernetes): Propertly handle null result from get and add log
Calling get() returns null when no resource is found rather than
throwing an exception. We should return Optional.empty() in that
case (rather than try to dereference the null in the following
logic) just as we do when we run into an exception.

Also add visibility by logging any exceptions in the exception block.