fix(kubernetes): Fix live lookup of CRDs
The logic in an earlier commit to look up CRD properties was
incorrect as it was looking up CRDs via the command
kubectl get crd kind.apigroup; we need to use the *plural* version
of the kind in this command (as a CRD's name is based on the plural
kind).

There's unfortunately no good way to directly look up a CRD based
on the singular name. This commit works around it by just listing
all CRDs and filtering based on the name to the one we're interested
in. To avoid the performance implications of too many lookups of
the entire list of CRDs, also re-introduce memoizing.