perf(kubernetes): Reduce namespace lookups in a loop
The call to primaryKinds() on a caching agent can be expensive,
as it may (depending on the caching agent) need to look up available
CRDs and/or available namespaces. We make this call in a tight loop
in a few places; pull the call out of the loop.

Also replace the lambda in liveNamespaceSupplier with an actual
private function for readability (and to make later refactoring easier).

Finally, we make a redundant call to credentials::isValidKind in the
CRD caching agent; primaryKinds() already filters on this before
returning.