refactor(kubernetes): Pull api versions to a set
A number of different places check the API version of a manifest
against a set of supported ones using a log if condition. Replace
these with a statically-defined set of supported versions and a
check of whether the manifest's API version is in the set.

This commit doesn't change any of the supported versions.