fix(kubernetes): Fix missing API versions (#4057)
This change cherry-picks the safe parts of #4055 and #4056
into the 1.16 branch. In particular, it pulls just the
additive changes that admit new API versions: we now accept
v1 Deployments and ReplicaSets, as well as
networking.k8s.io/v1beta1 Ingresses.

This change does *not* update the code to start deserializing
all existing objects as these new types, as that does have
the potential to break existing functionality. It also does not
cherry-pick the minor refactoring to pull the supported versions
into sets and avoid the long chain of or conditions.

The interest of cherry-picking this is that these new API versions
have been supported since Kubernetes 1.9 so even users who have not
fully upgraded may still run into errors using them (and the erorrs
are not necessarily easy to debug, leading to timeouts on waiting
for manifests to stabilize). This will also unblock users who have
upgraded to Kubernetes 1.16 (and also help us hear sooner if there
are other issues).