fix(kubernetes): Fix matching of artifacts in deploy/patch stages
When determining whether the deploy and patch manifest stages bound
all required artifacts, the artifacts in the list of required artifacts
have an artifact account set while those in the list we're trying to
bind don't.

As the .equals function of Artifact includes the account in its
comparison, this means that we don't recognize the replaced artifacts
as the ones we expected to replace and fail the stage.

Trying to fix this more generally is riskier than I'd like to cherry
pick, so as a temporary fix, define a class ArtifactKey that contains
the fields we'd like to use when comparing artifacts and use that
class for the comparison.