refactor(core): Move RequestQueue to core
A number of controllers currently use a RequestQueue to help
balance requests from different applications. One such
controller is the kubernetes-only ManifestController that I'd
like to move out of clouddriver-web and into
clouddriver-kubernetes-v2 as it is causing a fair amount of
kubernetes-specific code to live in core.

In order to support controllers (such as this one) that live
in their respective modules still using RequestQueue, move the
RequestQueue logic to core.