refactor(gce): don't serialize AutoscalingPolicy
Serializing Google API types doesn't work right with Jackson. These
types are actually all Map objects, so Jackson does weird things with
them. In particular, it's okay to for them to reference another Google
API object, but if _that_ object references a third Google API type
(e.g. AutoscalingPolicy has a ScaleDownControl field, which has a
FixedOrPercent field) then exceptions will be thrown when
deserializing.

Fortunately (and maybe for this exact reason??) we already have a class
GoogleAutoscalingPolicy which mirrors exactly the fields in
AutoscalingPolicy, so use that one instead.

The only real point of this commit is to change
GoogleServerGroup#autoscalingPolicy from a (compute API)
AutoscalingPolicy object to a (spinnaker) GoogleAutoscalingPolicy one,
so that serialization works better.

Changes to support this:
* There was a method
GCEUtil#buildAutoscalingPolicyDescriptionFromAutoscalingPolicy which
converts from AutoscalingPolicy to GoogleAutoscalingPolicy. Since we're
not storing AutoscalingPolicy anymore, we can get rid of those
conversions and just use the GoogleAutoscalingPolicy directly.
* But when we create a GoogleServerGroup in the caching agent, we still
need to do that conversion. I created a method in
GoogleAutoscalingPolicy to do it instead. I think this is a better place
than GCEUtil. (When I wrote it, I didn't realize one already existed in
GCEUtil. I think mine is better, anyway. And it has (some) tests!)