fix(kubernetes): Remove duplicate test assertions
I had intended the second set of assertions to be testing the
contents of the cache. But it loadData doesn't actually
update the cache (it just returns results to an outer function that
stores the result) so we can't test that the cache is updated after
calling loadData. Just remove the extra assertions.