fix(kubernetes): Add missing limitrange kind
Deployments containing a limitrange time out waiting for the
manifest to appear, as clouddriver is missing the entry for
this kind.

Ideally there would be a better way of auto-detecting kinds like
this (particularly ones that don't need any custom logic) but for
now just adding another Handler to handle this.