refactor(docker): Remove unused getTags() function
The function getTags() in DockerRegistryNamedAccountCredentials
has a bug. Depending on whether we return from the try or the
catch block, we'll get either an Instant or a Date, which we
then put into a single array and try to sort. In some cases
this can lead to an error 'Comparison method violates its general
contract' when we try to sort this heterogenous array.

While fixing this, I realized that this function is completely
unused since #3809, so just removing the function instead of fixing it.