fix(docker): Fix sorting array of non-comparable types
getCreationDate returns an Instant; when the call fails, we're
returning a default value as a Date. This means we can end up with
an array with a mix of Instant and Date objects, which don't
together obey the general Comparable contract.

Fix this by returning an Instant from the catch block; also pull
the error handling logic into a CompileStatic function so we
get some additional type safety.