refactor(docker): Pull key-based sorter into its own class
When sorting docker tags by creation date, we are effectively doing
a Schwartzian transform to map each item to a key, sort by that key,
then re-map back to the original item. Rather than have this ad-hoc
in Groovy, make a simple Java class that handles this algorithm and
that can be tested. This also allows us to mark the method that
contained this logic as CompileStatic, as we've removed the non-safe
operations into Java (and made them safe).

Right now, leaving the class alongside the single place it is used,
but if we need this in other places we can move it to a more general
location.