fix(titus): shutdown the executor for titus streaming agent
We are leaking an executor (with a single thread in) from the TitusStreamingAgent because it's not shutdown properly
I think it would be better to refactor this code to not create the executor (at least not all the time), but I want to make mininal change to address the immediate problem first.

There is also some unnamed thread pool that appears to be leaking that I can't track down. Perhaps it's some API that the streaming agent calls, will debug it more after this change goes in