test(kubernetes): Add tests to stateful/daemon set stability
The next commit will fix bugs in the stateful set and daemon
set stability conditions. Add tests to cover the cases that
will be fixed; in the case of the daemon set handler, there
were no existing tests so this commit adds some basic tests
in addition to tests of the bug that will be fixed.

This commit also does some minor cleanup on the stateful set
tests, removing some unused variables and referring to member
variables by their name instead of via their groovy-generated
accessors.