fix(kubernetes): Fix stability for stateful/daemon set
When scaling a stateful set down to zero replicas, we never
consider the stateful set stable. This is because the status
field for a stateful set with no replicas omits some fields
(instead of setting them to 0) and our logic does not properly
handle the null case. Fix this and add some tests to validate
the fix.

A similar bug exists with daemon sets; for daemon sets, you
don't explicitly set replicas but if you configure the daemon
set such that it runs on 0 nodes (for example by including a node
selector that doesn't match anything), Spinnaker will never
consider the daemon set stable.