fix(eureka): make non-404 error codes non fatal on disable ops
There already are a lot of special cases in this class, but marking the task as failed can have pretty bad consequences like failing entire deploy stages and pipelines, triggering rollbacks, etc. If the stars are really poorly aligned and we do get 10 failures (timeouts, 404s, 503s...) I think the kato task should still be considered a success, allowing orca to proceed to the next steps (most likely waiting for down instances) which have a chance of succeeding.