refactor(kubernetes): Remove isAccountRelevant and makesLiveCalls
Both implementations of KubernetesV2AbstractManifestProvider separately
call isAccountRelevant and makesLiveCalls to determine if an account
exists. This duplicates work and also adds an unnecessary layer of
indirection; instead juts have them get the optional account and
ask the account (if present) directly what they need to know.

This commit makes resourcePropertyResolver protected in the parent
class so that subclasses can use it, but a later commit will just
push it directly down to the subclasses.