refactor(kubernetes): Pass credentials instead of account name
In a number of the functions in the KubernetesManifestProvider (and
LiveManifestProvider) we pass around an account name even when we
already have the credentials; to avoid having the function we call
need to resolve the credentials again, just pass the credentials.

To account (no pun intended) for cases where we need the name of the
account (generally for logging), add a Getter to get the account name
from the credentials.

This allows us to remove getRegistry from the base manifest provider
as it is no longer used.