refactor(kubernetes): Eliminate abstract base class
At this point, all the KubernetesV2AbstractManifestProvider is
doing is supplying a single static method. Rather than using
inheritance to share it between the two implementations, we'll
use composition.

* Change the two concrete manifest providers to directly
implement ManifestProvider<KubernetesV2Manifest>
* Rename KubernetesV2AbstractManifestProvider to
KubernetesV2ManifestBuilder as it helps build a manifest
* Call the static method in KubernetesV2ManifestBuilder from
the manifest providers

The abstract class being eliminatd is not one that should be
used by people extending Spinnaker; the only reason there is a
base class at all was to sit on top of live- and non-live- manifest
providers.

There may be a better place for the single method in
KubernetesV2ManifestBuilder but for now I'll leave it there and
will leave finding a better place to a future refactor.