refactor(kubernetes): Support kind lookup directly on credentials
I'd originally allowed callers to get the KindRegistry associated
with a KubernetesV2Credentials object as it's a smaller thing to
pass around in cases where we don't need the credentials and I
didn't want to add more functionality to an already large class.

But the reality is that every time we look something up on a registry,
we do so via credentials.getKindRegistry().getSomething() so this really
is something that clients are querying on the credentials object.

By having clients just call getGlobalKinds and getKindProperties
directly on the credentials object, we make the existence of the
kind registries an implementation detail which makes them easier
to change as we account for more conditions and edge cases in
computing the status of a kind. (An upcoming commit will require
us to consider both the status in the registry and whether a kind
is readable when computing its status which is now much easier to
do without affecting callers.)