refactor(kubernetes): Make registries package-private
Now that the kind registries are implementation details and all
callers just directly query the credentials for information on
kinds of interest, make these package-private so they aren't used
in other places.

This also changes GlobalKubernetesKindRegistry to be a component
instead of being generated as a bean (which would have required it
to be visible to the configuration class).

Also add default nullity annotations to these registries.