refactor(kubernetes): Simplify some credentials functions
The only callers of getGlobalKinds() and getCrds() take the result
and:
* Discard the properties to get only the actual kind
* Filter out invalid kinds

Build this into the credentials, so that these functions just directly
return kinds, and only return valid ones, and remove the filtering
logic from the callers.

This also allows us to make isValidKind private as we're doing filtering
within the class.

Also update getRegisteredKinds in the kind registries to more directly
return what the only caller is now asking for.