refactor(kubernetes): Add isRegistered function to kind registry
There's currently no way for a consumer of the KubernetesKindRegistry
to know if a kind is registered as getKindProperties will return default
properties if it isn't.

Address this in two ways:
(1) Rename getKindProperties to getKindPropertiesOrDefault to make it
more clear what this function does.
(2) Add an isKindRegistered that will query if a kind is registered (and
try to register it by looking up CRD properties) and returns true if
the kind is registered at the end of the function (either initially or
by CRD lookup).