fix(kubernetes): Allow deploying a CRD and object in same manifest
Currently a multi-manifest that deploys a CRD and an object referencing
that CRD will fail, because we validate that all objects are of a
kind that exists in the cluster, but the CRD won't exist until the
manifest is deployed.

Let's fix this by adding a kind status UNKNOWN, which means that
Spinnaker does not know about this kind (and can't find it in the
cluster's CRDs) and allow deploying manifests whose kind has this
status.