test(kubernetes): Add tests to status function of handlers
There are very few tests on the implementations of the status
function in KubernetesHandler. Before doing some refactoring of
the status computation logic, add a robust suite of tests to
document the current behavior.

Some notes:
* There were already tests on daemon set and stateful set; here
I've augmented the tests to assert more thing on the returned
status, and in the case of stateful set have converted the groovy
test to Java.
* Some of these might be a bit overspecified (ex: checking the
specific string in the unstable reason) but I'd rather know if
something changes during the refactor (and may make some slight
deliberate changes).
* I tried as much as possible to put realistic YAML in the resources
(ex: by doing a rollout and watching the status in my cluster). In
some cases this led me to suspect that we are testing conditions that
could never be true; I added comments in places where this is the case.