fix(kubernetes): Don't return early after setting failed flag
When a deployment reports that it is failed we exit early instead
of checking the remaining condition. This means that we don't always
properly set the other flags of the status in this case; update the
logic to set the other flags.

The logic in orca to decide how to handle the status of a manifest
checks both the stable and the failed flags (so we won't risk breaking
it by sometimes setting a deployment to unstable when it is also
failed, which we already do in some cases anyway if the condition
reports unavailable).