refactor(kubernetes): Improve Status and Condition classes
There are way too many ways to create and mutate these classes,
as both have fully public fields as well as setters; in the case
of Status there are also custom setters, and in the case of Condition
there is additionally a builder.

Clean this up as follows:
* Update the custom setters (ex: failed, unstable, etc.) in Status
to create a new condition and assign it to the appropriate value
instead of mutating the existing condition. (This fixes a null-safety
issue as some of these fields can be null, if unknown() has been
called.)
* This now allows us to make Condition immutable, as we never change an
existing condition. Make it immutable by removing the setters and
marking the fields final. Replace builders with a constructor (which is
reasonable given that we have only two fields with different types). As
it's common to create a condition with a null message (because it shows
up in defaults for Status) make a convenience factory method withState
that returns a Condition with null message.  (Because the class is
immutable we can just return a pre-created static Condition in this
case.)
* Add a defaultStatus() factory method to Status; it's the same as the
constructor new Status() but I think makes the intent much more clear
when reading the Handlers; that is, start with some default status and
mutate it if the desired status differs from the default.  I didn't make
the no-arg constructor private because I didn't want to needlessly break
people who used it in their own custom handlers, but ideally they'd
migrate to defaultStatus() for clarity.
* Add a stable() mutator to replace the one case where we explicitly
called setStable()
* Add nullity annotations.  Many things are nullable; the new
annotations just document that existing behavior.