refactor(kubernetes): Simplify status reporting in handlers
The logic to report workload status is fairly complex to follow;
one of the reasons it's complex is that in general we create an
empty Status object (which defaults some fields) at the beginning
of the status function and mutate it throughout the functions,
sometimes returning immediately after the mutation but sometimes not.

In order to understand what we're returning at any given point, we
need to read the entire function up to that point to see if any
of the prior conditions wrote to the result variable.

Where possible, simplify this by just creating a new Status inline
at the point of return (and change from using new Status() to the
factory method Status.defaultStatus()). This is possible in any
cases where there are no mutations of result earlier in the function
that don't immediately return. By doing this analysis now once, we
make it easier for future readers of the function to see what is
returned in each case.

This commit only touches the handlers where it was trivial to make
this change (ie, we never mutate the result without immediately
returning it).