fix(kubernetes): Fix ready condition check for replicaset
We currently check the available replicas count before we check
the ready replica count; this isn't correct as there will always
be more ready replicas than available replicas. (The definition
of an available replica is one that has been ready for at least
minReadySeconds.) Switch the order of the check so that we
generate the correct message in each case.