fix(kubernetes): Simplify status messages
When a replicaset is unstable and unavailable we use slightly
different wording in the unstable and unavailable message even
though the cause is the same. This isn't complicates both the
code and the user's parsing of messages (as they need to parse
both to realize they are the same error).

Simplify this; the tests show exactly where the text of errors has
changed and shows that the actual stable/unstable status does not
change.

Also, push the Deployment's UnstableReason up to the top level so we
can re-use it for replica sets.