fix(kubernetes): Replace null status with false
When the observed generation in the status does not match the object's
generation, we set the stable and failed conditions to null. This should
not be treated any differently than if the status generation did match
but the object had not fully rolled out---it just means we need to wait
a bit longer to see the new status.

The recent changes to orca have explicitly made it so that from orca's
perspective there is no different beteween stable/failed being null and
stable/failed being false. Make things more clear (and null-safe) by
removing the ability to null these conditions.