fix(kubernetes): Move generation checks to be object-specific
We have a generic function in KubernetesManifest to determine
whether the status generation is correct. It has a lot of type-safety
and null-safety issues as it's operating on an untyped map.

As each status handler already converts the manifest to a strongly-typed
representation of its handled object, let's move the status checks to
happen on that object.

This in particular fixes some null-pointer issues that were documented
in the tests.