refactor(kubernetes): Make condition constructors private
As I was thinking about how these changes might affect users with
custom handlers, I realized that there's actually nothing anyone
can do with a Condition directly anymore.  Since there are no
setters on Status, you can't pass a condition to a status; it's
up to Status to create this based on your message.

Based on this, make the Condition constructor/factory private to
simplify the API. This doesn't break anything additional as there's
nothing you could do with this condition.

Also making the class final (as making the constructor private
effectively does this).