refactor(kubernetes): Push logic for kubernetes health to base
The logic for checking the health of a Kubernetes account is similar
to what we want for docker accounts (and likely for other accounts).
Namely, we want to report errors with an individual account, but don't
want to bring down all of clouddriver because of failures with
individual accounts. Make a base class that can be extended by cloud
providers to implement this functionality.

This also makes one small change to the way unhealthy accounts are
reported in detailed health check results; as the accounts will already
appear under a key of 'kubernetes', I've eliminated the redundant prefix of
'kubernetes:' from each of the account names.