fix(docker): Fix docker registry health check
The docker registry health check currently fails if any single docker
account is unreachable.  (Though it will not fail after startup if it
has passed once.)

This means that a single bad account can cause all of clouddriver to
fail to start, which is bad for failure isolation. Rewrite the health
check to use the new AccountHealthIndicator class which will report
errors (both as metrics and in the detailed health view) but will not
cause clouddriver to report itself as DOWN.