fix(kubernetes): Fix tests depending on map iteration order
When applied to a Map, containsExactly requires that iterating over
the map returns the supplied entries in order. As we are not using
an ordered map, this will lead to intermittent test failures.

Replace this with containsOnly, which asserts the map contains
exactly the supplied entries, in any order.

Also fix a copyright header.