refactor(artifacts): Compile clouddriver-artifacts with Java compiler
This package is already entirely in java, so no additional changes
are needed other than setting the flag to compile it with the java
compiler.

We can also remove the useJUnitPlatform() setting as it has now been
added to the base clouddriver build file so does not need to be
specified in individual modules.