refactor(artifacts): Make artifact classes final and package-private
In general, the interface of the artifacts module is to allow clients
to pass in an artifact and get back an InputStream of the resolved
artifact's contents. All of the accounts, configuration, and credentials
are just implementation details of how this module does that and should
be made package-private (and final as they were not designed for
inheritance).

The interfaces of course remain public, but clients should not need
to know which particular implementation of an artifact account is
handling their request.  In general, this is the case, though there
is one place that breaks this encapsulation, where the cloud foundry
deploy depends on details of the Maven artifact resolver. For that
reason (to keep this refactor smaller) I'm just leaving that resolver
public but ideally we'd find a better way of supporting what that
consumer needs rather than make the entire resolver public.