fix(redis): Clean up cache hashing config
The config setting caching.redis.hashingEnabled controls whether
we try to short-circuit sending data to redis by comparing a hash
of the data client-side with a hash of the data server-side.

This setting is on by default, (see DEFAULT_HASHING_ENABLED in
RedisCacheOptions.java) so there's no need to explicitly set it
in config files.

Also remove the setting turning it off in the HA clouddriver instances;
this has no effect for -ro and -ro-deck as those instances never
write to the cache anyway. This will be a change for the -rw instance,
but in general this should improve performance. (At least there's no
reason a priori to have the default different between different
clouddriver instances.)