refactor(core): Remove explicit setting of shutdownWaitSeconds
This parameter controls how long clouddriver will wait for in-flight
tasks before shutting down. It is defaulted to 600 for all users
deploying with halyard (which is most users). Rather than put this
setting directly in the halyard-generated config, set the default
in code.

This will change the default for users that don't use Halyard
(primarily Netflix). I'm happy to pick a different default if that
is more reasonable, but given this is what all Halyard users have
had for years it seems like the smallest change.