refactor(core): Push defaulting of port and host to base config
Halyard has been using the settings in spinnaker.yml to default
the port and host for clouddriver for all Halyard users by adding
a block to the user's config.

Push this logic to the base clouddriver config; the defaults are
the same for anyone that does not have a services.* block in their
config overriding the values, but for users who do (all Halyard users)
this now respects those overrides.